LayoutTests/imported/w3c/web-platform-tests/speculation-rules/prefetch/different-initiators.sub.https_cross-site-2-expected.txt


FAIL Cross-initiator prefetches using ServiceWorker tricks assert_equals: Prefetch should not work due to ServiceWorker. expected (undefined) undefined but got (string) "prefetch"
